New jquery editor plugin.
The new editor plugin does not work and is a work in progress, as a result
a new config has been added and the editor can be shown/hidden via a config
option found in the config builder.

As it does not work, do not use it, thus disabled by default until it is working.